fix: resolve clippy lint violations in PR #89
- Replace manual range checking with RangeInclusive::contains (task_aging.rs:186)
- Remove unnecessary f64 to f64 cast in variance calculation (fairness_integration.rs:273)

All tests pass with RUSTFLAGS="-D warnings" and no clippy violations.

+++ crates/core/tests/fairness_integration.rs
@@ -270,7 +270,7 @@ fn fairness_multi_round_consistency() {
     let avg = round_stats.iter().sum::<usize>() as f64 / round_stats.len() as f64;
     let variance = round_stats
         .iter()
-        .map(|&x| ((x as f64 - avg).powi(2)) as f64)
+        .map(|&x| (x as f64 - avg).powi(2))
         .sum::<f64>()
         / round_stats.len() as f64;
